requirements.txt
# Core data stack
pandas>=2.0
numpy>=1.25

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Modeling
scikit-learn>=1.3
xgboost>=1.7

# Experiment tracking
mlflow>=2.12

# Dashboard
streamlit>=1.25

# Date handling
python-dateutil>=2.9

# API framework
fastapi>=0.100
uvicorn[standard]>=0.35

# NLP & Embeddings
sentence-transformers>=2.2
faiss-cpu>=1.7

# PDF generation
fpdf>=1.7

# Environment config
python-dotenv>=1.0

# OpenAI-compatible client
openai>=1.3
